+++ src/LDPC/NBLLR.cpp
@@ -6,14 +6,14 @@
  * @param d data
  * @param q Q, should in [1, GF]
  */
-NBLLR::NBLLR(double d, int q) {
+NBLLR::NBLLR(float d, int q) {
     data = d;
     Q = q;
 }
 
 NBLLR::~NBLLR() {}
 
-double NBLLR::getLLR() const {
+float NBLLR::getLLR() const {
     return data;
 }
